ISPN-783 - In Memcached, connections to be closed only when SERVER_ERROR
Basically, those client errors that should not close connections need to
be propagated via a different path rather than allowing the decode()
call to throw an exception. This allows for pipelined requests after
client errors to be processed correctly.

Also tidied up a lot of the exception handling code to avoid code
duplication when figuring out whether an exception is a client error or
not, and when building the actual error response/message back.