ISPN-930 - race condition in Lock reordering could cause Lucene Index corruption when a CacheLoader is enabled
The workaround is to avoid writing two different values under the same key.
With Lucene, the first bytes are rewritten at file close time to add checksum
information to the head of the file.
Previously we where writing each file chunk as soon as possible to be
memory friendly, getting the first chunk back as needed to edit it and putting
it into the cache modified - apparently if these operations are close enough in
time, the value without the checksum might end up to be the value stored (ISPN-575).
What we do with this patch is hold the writing of the first chunk up to the file
close operation, so that each chunk is written only once.

As a side effect, we can simplify the batch operations and the intermediate flush
operations, as they don't need to detect if we are in closing phase.