ISPN-870 Virtual Nodes
Initial implementation of virtual nodes implemented
on top of the topology aware consistent hash. It
can be enabled by setting the numVirtualNodes
configuration parameter on clustering.hash

The implementation of virtual nodes passes existing
tests verifying that the hash function is working
however no tests have been introduced to test the
correctness of the virtual node distribution.

The implementation is entirely self contained in
the consistent hash function, and the bulk of it
is provided in AbstractWheelConsistentHash, which
contains instructions for implementations of a
wheel consistent hash to enable virtual nodes.