ISPN-1123 - Extend the TCPPING port_range to make the initial member discovery more reliable
The old port_range was 2, so if the first two ports were busy the coordinator would not be in the port range and a second node wouldn't find it.
I changed it to 30 to match the TCP port_range.

I also moved waitForRehashToComplete/waitForInitRehash to TestingUtil and I modified most tests to use MultipleCacheManagersTest.waitForClusterToComplete instead.