ISPN-1123 - Revert TCPPING port_range change
A very large port_range value seems to lead to timeouts waiting for the cluster to form.
Seems to be especially bad with high CPU load, as I only got the timeouts on my machine when I has MapStressTest enabled.
So I changed the port_range back from 30 to 3 and I disabled the CH performance tests.