ISPN-1194 - Consolidate rehash and replication state retrieval code
Moved the rehashing logic out of the DistributionManager to a new class DistributedStateTransferManager.
Extracted the transaction lock out of the TransactionLogger into a new class TransactionLock.
Removed the dependency on a coordinator, using instead a broadcast to signal to all nodes that push is complete.
Added a new REQUEST_JOIN DistributedStateTransferControlCommand to make sure joiners are properly started before the old members start sending any commands. The user will still need to start all the caches upfront, as we don't support asymmetric caches.