ISPN-2281 Hot Rod uses raw values with Cache taking versions
* AdvancedCache adds a couple of new methods that take metadata
as parameter. This can be used by users to provide version, and
also for server implementations to add their own metadata (i.e.
REST servers and MIME data).
* The implementation of these methods is temporary and is bound
to change in order to feed the metadata object all the way to
the cache entry.
* There's an initial stab at the compatibility testing with
initial focus on Embedded and Hot Rod server. Tests here are not
yet working.