ISPN-3318 Automated upgradable path for file store consumers
* Remove SingleFileCacheStore and related files and ported the code
over to FileCacheStore.
* Old FileCacheStore logic moved to BucketFileCacheStore which is
package private in order to avoid instantiation.
* FileCacheStore can now detect old bucket-based data structure and
it can upgrade it, providing a backup for the old data.
* Backup is still accessible, by enabling the deprecatedBucketFormat
configuration attribute, in which case the old bucket-based
FileCacheStore is used.
* In order to avoid having to replicate the configuration hierarchy
for the old, bucket-based cache store, created a delegate cache store
which delegates to the correct cache file cache store (new, or bucket
based), depending on whether the deprecatedBucketFormat flag was
enabled or not.
* This is not ideal and in fact I suggest some alternative solution
as comment in the code.