ISPN-3400 Upgrade to JSR-107 (JCache) 1.0.0-PFD version
* Upgraded to Public Final Draft version of the JSR-107 specification.
* TCK has expanded loadAll() and expiry policy usage tests, which in
turn has required more complex logic to deal with all the expectations
in the TCK. For example, the logic for loadAll() is now different based
on whether a JCache cache loader has been configured, or the Infinispan
configuration has a native cache loader configured.
* Tweaked a test util method to provide more generic friendly returns
that avoids the need for cast.