ISPN-3316 Avoid use of @Alternative for Infinispan components
* In CDI 1.0, alternative components cannot be injected in libraries
used by deployments. In other words, only components used by the app
directly can be switched with alternative.
* An alternative set of Cache*Interceptor classes have been developed
for use in managed environments, which indirectly use the
InjectedCacheResolver.
* Indirectly solved an issue where cache managers where being left
open in CDI/JCache integration tests. This was the result of
JCacheManager trying to stop cache managers for which it didn't
control lifecycle.
* Remove license headers from all source files that had them.