ISPN-4546 Fix retry for write commands in a transaction
* Retry write commands (and generated LockControlCommands).
TODO Retrying a write adds a duplicate command in the modifications list.

* Don't retry a tx command after it has succeeded on the originator
For a commit command, if there are new owners, they will receive the
updated values via state transfer. On the other hand, a retried commit
may not find the remote transaction on the new owners.
For lock and prepare commands, the new owners will also receive the
transaction via state transfer.