ISPN-5429 Avoid CacheEntryFilterConverter duplicate callbacks
* When plugging a CacheEntryFilterConverter in a cluster, it can lead
to duplicate callbacks as a result of listener invocation base class
not detecting that the CacheEntryFilterConverter is both a filter
and converter.
* When clustering CacheEntryFilterConverter instances, these are
wrapped to avoid pre-events, and the wrapping halts the
discovery that the CacheEntryFilterConverter is both a filter and
converter.
* Overriding equals() in wrapper classes and making sure base listener
invocation uses equals() instead of referential equality solves the
problem.
* Added remote listener tests to verify the expected amount of
callbacks are received by CacheEntryFilterConverter implementations.